requirements.txt
pwntools==4.9.0
requests==2.29.0
pymongo==4.3.3